ensure elasticsearch is running
Check if systemd unit is running or not. If it is not running start
it.
Note: it may take some time for the service to be available, because at the
beginning much data is processed. Maybe introduce some waiting after
the start to make sure the servie really is available.