make a class for the search
This makes some parts easier. Further improvements are much better to
include now. Also I got rid of the global variables, what is a value
for itself.

Note: This is not yet finished. Some parts are still to be
improved. For example the printing is at this moment not aware of the
terminal size (don't know wether this is important or not).