remove automatic start of Elasticsearch
Do not start the Elasticsearch service anymore. The user has to do
this on his own.
It is not clear whether everybody uses systemd to start the service or
not, so everybody has to choose the right method of starting the
service.
In addition urllib3 will throw an error because Elasticsearch is not
online fast enough and will not accept connections.